render job now creates files

+++ application/jobs/Render.php
@@ -23,26 +23,14 @@ class Render extends \Flexio\Jobs\Base
         $this->getOutput()->setEnv($this->getInput()->getEnv());
         $input = $this->getInput()->getStreams();
 
-        foreach ($input as $instream)
-        {
-            $this->createOutput($instream);
-        }
-    }
-
-    private function createOutput(\Flexio\Object\Stream $instream)
-    {
         $job_definition = $this->getProperties();
-        $mime_type = $instream->getMimeType();
-
         $url = $job_definition['params']['url'] ?? null;
         $format = $job_definition['params']['format'] ?? 'pdf';
+        $width = $job_definition['params']['width'] ?? null;
+        $height = $job_definition['params']['height'] ?? null;
 
         if (!isset($url))
             throw new \Flexio\Base\Exception(\Flexio\Base\Error::MISSING_PARAMETER);
-
-        $dockerbin = \Flexio\System\System::getBinaryPath('docker');
-        if (is_null($dockerbin))
-            throw new \Flexio\Base\Exception(\Flexio\Base\Error::INVALID_PARAMETER);
 
         if ($format == 'pdf')
             $content_type = 'application/pdf';
@@ -51,9 +39,22 @@ class Render extends \Flexio\Jobs\Base
         else
             throw new \Flexio\Base\Exception(\Flexio\Base\Error::INVALID_PARAMETER);
 
-        $outstream = $instream->copy()->setPath(\Flexio\Base\Util::generateHandle())->setMimeType($content_type);
+        // get docker binary
+        $dockerbin = \Flexio\System\System::getBinaryPath('docker');
+        if (is_null($dockerbin))
+            throw new \Flexio\Base\Exception(\Flexio\Base\Error::INVALID_PARAMETER);
+
+
+        // create the output stream
+        $outstream_properties = array(
+            'name' => $name,
+            'mime_type' => $content_type
+        );
+        $outstream = \Flexio\Object\Stream::create($outstream_properties);
         $this->getOutput()->addStream($outstream);
+
         $streamwriter = \Flexio\Object\StreamWriter::create($outstream);
+
 
         if ($format == 'pdf')
             $cmd = "$dockerbin run -a stdin -a stdout -a stderr --rm -i fxchrome sh -c 'timeout 30s google-chrome --headless --disable-gpu --print-to-pdf --no-sandbox $url && cat output.pdf'";
